+++ migrations/versions/98c3d8971d6c_first_update.py
@@ -1,4 +1,4 @@
-"""First Update.
+"""First Update - Add elastic, sshkey cols to deployments, add users fk
 
 Revision ID: 98c3d8971d6c
 Revises: 88bc3c2c02a6
@@ -7,7 +7,7 @@ Create Date: 2020-02-28 11:18:05.926021
 """
 from alembic import op
 import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
+from app.lib import alembic_helper
 
 # revision identifiers, used by Alembic.
 revision = '98c3d8971d6c'
@@ -17,14 +17,20 @@ depends_on = None
 
 
 def upgrade():
-    op.add_column('deployments', sa.Column('elastic', sa.Boolean, server_default=sa.sql.False_(), nullable=False))
-    op.create_foreign_key(None, 'deployments', 'users', ['sub'], ['sub'])
-    op.add_column('users', sa.Column('sshkey', sa.Text(), nullable=True))
+    if not alembic_helper.column_exists('deployments', 'elastic'):
+        op.add_column('deployments', sa.Column('elastic', sa.Boolean, server_default=sa.sql.False_(), nullable=False))
+    if not alembic_helper.fk_exists('deployments', 'users', ['sub'], ['sub']):
+        op.create_foreign_key(None, 'deployments', 'users', ['sub'], ['sub'])
+    if not alembic_helper.column_exists('deployments', 'sshkey'):
+        op.add_column('users', sa.Column('sshkey', sa.Text(), nullable=True))
     # ### end Alembic commands ###
 
 
 def downgrade():
-    op.drop_column('users', 'sshkey')
-    op.drop_constraint(None, 'deployments', type_='foreignkey')
-    op.drop_column('deployments', 'elastic')
+    if alembic_helper.column_exists('deployments', 'sshkey'):
+        op.drop_column('users', 'sshkey')
+    if alembic_helper.fk_exists('deployments', 'users', ['sub'], ['sub']):
+        op.drop_constraint(None, 'deployments', type_='foreignkey')
+    if alembic_helper.column_exists('deployments', 'elastic'):
+        op.drop_column('deployments', 'elastic')
     # ### end Alembic commands ###
